pyproject.toml
[tool.poetry]
name = "bulletin_board"
version = "0.1.0"
description = ""
authors = ["Artem <artem.mikhaylov5593@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
pandas = "^1.4.3"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
drf-writable-nested = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
